Add files via upload
This is an Microsoft Office Excel file with a Macro (Visual basic) that can be used to define sprites. It can create standard basic code e.g. defining the data and appropriate code to read and poke the definition in memory. In this repository it is placed to create the chr$ values to define a string that can be used to print the sprint values.